@@ -38,7 +38,7 @@ public class Chord implements AbcConstants
 	private long startTick;
 	private long endTick;
 	private boolean hasTooManyNotes = false;
-	private List<NoteEvent> notes = new ArrayList<NoteEvent>();
+	private List<NoteEvent> notes = new ArrayList<>();
 	private int highest = 0;
 	private int lowest = 200;
 
@@ -199,204 +199,198 @@ public class Chord implements AbcConstants
 
 	public List<NoteEvent> prune(final boolean sustained, final boolean drum) {
 		// Determine which notes to prune to remain with a max of 6
-		List<NoteEvent> deadNotes = new ArrayList<NoteEvent>();
+		List<NoteEvent> deadNotes = new ArrayList<>();
 		if (size() > MAX_CHORD_NOTES) {
 			recalcEdges();
 
-			List<NoteEvent> newNotes = new ArrayList<NoteEvent>();
+			List<NoteEvent> newNotes = new ArrayList<>();
 			
-			Comparator<NoteEvent> keepMe = new Comparator<NoteEvent>() {
+			Comparator<NoteEvent> keepMe = (n1, n2) -> {
 
-				@Override
-				public int compare(NoteEvent n1, NoteEvent n2) {
-					
-					if (n1.note == Note.REST) {
-						return 1;
+				if (n1.note == Note.REST) {
+					return 1;
+				}
+				if (n2.note == Note.REST) {
+					return -1;
+				}
+
+				/*
+				if (n1.doubledNote && !n2.doubledNote) {
+					return -1;
+				}
+				if (n2.doubledNote && !n1.doubledNote) {
+					return 1;
+				}*/
+
+				// At the point in time when prune() is run, there is very few tiedTo, mostly just tiedFrom.
+				//assert n1.tiesTo == null;
+				//assert n2.tiesTo == null;
+				/* This was commented as experiments in lotro shows than can start a new note even if 6 prev. is still playing.
+				boolean n1Finished = false;
+				boolean n2Finished = false;
+				if (!sustained) {
+					// Lets find out if the notes have already finished
+					long dura = 0;
+					for (NoteEvent neTie = n1.tiesFrom; neTie != null; neTie = neTie.tiesFrom)
+					{
+						dura += neTie.getLengthMicros();
 					}
-					if (n2.note == Note.REST) {
+					if (dura > AbcConstants.NON_SUSTAINED_NOTE_HOLD_SECONDS) {
+						n1Finished = true;
+					}
+					dura = 0;
+					for (NoteEvent neTie = n2.tiesFrom; neTie != null; neTie = neTie.tiesFrom)
+					{
+						dura += neTie.getLengthMicros();
+					}
+					if (dura > AbcConstants.NON_SUSTAINED_NOTE_HOLD_SECONDS) {
+						n2Finished = true;
+					}
+				}
+				if (n1Finished && !n2Finished) {
+					return -1;
+				} else if (n2Finished && !n1Finished) {
+					return 1;
+				}
+				*/
+
+				if (!sustained) {
+					// discard tiedFrom notes.
+					// Although we already checked for finished notes,
+					// we don't mind stopping note and not let it decay
+					// to prioritize a new sound.
+					if (n1.tiesFrom != null && n2.tiesFrom == null) {
 						return -1;
 					}
-					
-					/*
-					if (n1.doubledNote && !n2.doubledNote) {
-						return -1;
-					}
-					if (n2.doubledNote && !n1.doubledNote) {
-						return 1;
-					}*/
-										
-					// At the point in time when prune() is run, there is very few tiedTo, mostly just tiedFrom.
-					//assert n1.tiesTo == null;
-					//assert n2.tiesTo == null;
-					/* This was commented as experiments in lotro shows than can start a new note even if 6 prev. is still playing.
-					boolean n1Finished = false;
-					boolean n2Finished = false;
-					if (!sustained) {
-						// Lets find out if the notes have already finished
-						long dura = 0;
-						for (NoteEvent neTie = n1.tiesFrom; neTie != null; neTie = neTie.tiesFrom)
-						{
-							dura += neTie.getLengthMicros();
-						}
-						if (dura > AbcConstants.NON_SUSTAINED_NOTE_HOLD_SECONDS) {
-							n1Finished = true;
-						}
-						dura = 0;
-						for (NoteEvent neTie = n2.tiesFrom; neTie != null; neTie = neTie.tiesFrom)
-						{
-							dura += neTie.getLengthMicros();
-						}
-						if (dura > AbcConstants.NON_SUSTAINED_NOTE_HOLD_SECONDS) {
-							n2Finished = true;
-						}
-					}
-					if (n1Finished && !n2Finished) {
-						return -1;
-					} else if (n2Finished && !n1Finished) {
+					if (n2.tiesFrom != null && n1.tiesFrom == null) {
 						return 1;
 					}
-					*/
-					
-					if (!sustained) {
-						// discard tiedFrom notes.
-						// Although we already checked for finished notes,
-						// we don't mind stopping note and not let it decay
-						// to prioritize a new sound.
-						if (n1.tiesFrom != null && n2.tiesFrom == null) {
-							return -1;
-						}
-						if (n2.tiesFrom != null && n1.tiesFrom == null) {
+				}
+
+
+				if (n1.velocity != n2.velocity) {
+					// The notes differ in volume, return the loudest
+					return n1.velocity - n2.velocity;
+				}
+
+				if (!drum) {
+					if (n1.note.id != n2.note.id) {
+						// return the note if its the highest in the chord
+						if ((n1.origPitch == 0 && highest == n1.note.id)||(n1.origPitch != 0 && highest == n1.origPitch)) {
 							return 1;
 						}
-					}
-					
-					
-					if (n1.velocity != n2.velocity) {
-						// The notes differ in volume, return the loudest
-						return n1.velocity - n2.velocity;
-					}
-										
-					if (!drum) {
-						if (n1.note.id != n2.note.id) {
-							// return the note if its the highest in the chord
-							if ((n1.origPitch == 0 && highest == n1.note.id)||(n1.origPitch != 0 && highest == n1.origPitch)) {
-								return 1;
-							}
-							if ((n2.origPitch == 0 && highest == n2.note.id)||(n2.origPitch != 0 && highest == n2.origPitch)) {
-								return -1;
-							}
-							// return the note if its the lowest in the chord
-							if ((n1.origPitch == 0 && lowest == n1.note.id)||(n1.origPitch != 0 && lowest == n1.origPitch)) {
-								return 1;
-							}
-							if ((n2.origPitch == 0 && lowest == n2.note.id)||(n2.origPitch != 0 && lowest == n2.origPitch)) {
-								return -1;
-							}
+						if ((n2.origPitch == 0 && highest == n2.note.id)||(n2.origPitch != 0 && highest == n2.origPitch)) {
+							return -1;
 						}
-						
-						if (n1.note.id == n2.note.id) {
-							// The notes have same pitch and same volume. Return the longest.
-							// The code should not get in here.
-							return (int) (n1.getFullLengthTicks() - n2.getFullLengthTicks());
+						// return the note if its the lowest in the chord
+						if ((n1.origPitch == 0 && lowest == n1.note.id)||(n1.origPitch != 0 && lowest == n1.origPitch)) {
+							return 1;
 						}
-						
-						int points = 0;
-						
-						List<Integer> removeFirst = new ArrayList<Integer>();
-						
-						removeFirst.add(highest-32);
-						removeFirst.add(highest-24);
-						removeFirst.add(highest-12);
-						removeFirst.add(lowest+32);
-						removeFirst.add(lowest+24);
-						removeFirst.add(lowest+12);
-						
-						int index1 = removeFirst.indexOf(n1.note.id);
-						int index2 = removeFirst.indexOf(n2.note.id);
-						
-						// Discard notes first that has octave spacing from highest or lowest notes
-						if (index1 != -1) {
+						if ((n2.origPitch == 0 && lowest == n2.note.id)||(n2.origPitch != 0 && lowest == n2.origPitch)) {
+							return -1;
+						}
+					}
+
+					if (n1.note.id == n2.note.id) {
+						// The notes have same pitch and same volume. Return the longest.
+						// The code should not get in here.
+						return (int) (n1.getFullLengthTicks() - n2.getFullLengthTicks());
+					}
+
+					int points = 0;
+
+					List<Integer> removeFirst = new ArrayList<>();
+
+					removeFirst.add(highest-32);
+					removeFirst.add(highest-24);
+					removeFirst.add(highest-12);
+					removeFirst.add(lowest+32);
+					removeFirst.add(lowest+24);
+					removeFirst.add(lowest+12);
+
+					int index1 = removeFirst.indexOf(n1.note.id);
+					int index2 = removeFirst.indexOf(n2.note.id);
+
+					// Discard notes first that has octave spacing from highest or lowest notes
+					if (index1 != -1) {
+						points += -2;
+					}
+					if (index2 != -1) {
+						points += 2;
+					}
+
+					if (sustained) {
+						// We keep the longest note, including continuation from notes broken up
+						if (n1.getFullLengthTicks() + n1.continues > n2.getFullLengthTicks() + n2.continues) {
+							points += 2;
+						} else if (n2.getFullLengthTicks() + n2.continues > n1.getFullLengthTicks() + n1.continues) {
 							points += -2;
 						}
-						if (index2 != -1) {
-							points += 2;
-						} 
-						
-						if (sustained) {
-							// We keep the longest note, including continuation from notes broken up
-							if (n1.getFullLengthTicks() + n1.continues > n2.getFullLengthTicks() + n2.continues) {
-								points += 2;
-							} else if (n2.getFullLengthTicks() + n2.continues > n1.getFullLengthTicks() + n1.continues) {
-								points += -2;
-							}
-						}
-											
-						if ((Math.abs(n1.note.id - n2.note.id) == 12 || Math.abs(n1.note.id - n2.note.id) == 24 || Math.abs(n1.note.id - n2.note.id) == 36)) {
-							// If 2 notes have octave spacing, keep the highest pitched.
-							if (n1.note.id > n2.note.id) {
-								points += 2;
-							} else if (n2.note.id > n1.note.id) {
-								points += -2;
-							}
-						}
-						
-						if (sustained) {
-							if (n1.tiesFrom != null) {
-								points += 1;
-							}
-							if (n2.tiesFrom != null) {
-								points += -1;
-							}
-						}
-						
-						if (points > 0) return 1;
-						if (points < 0) return -1;
-						
-					} else {
-						
-						// Bass drums get priority:
-						if (n1.note == Note.As3) {// Open bass
-							return 1;
-						} else if (n2.note == Note.As3) {
-							return -1;
-						} else if (n1.note == Note.D3) {// Bass slap 2
-							return 1;
-						} else if (n2.note == Note.D3) {
-							return -1;
-						} else if (n1.note == Note.Gs3) {// Bass
-							return 1;
-						} else if (n2.note == Note.Gs3) {
-							return -1;
-						} else if (n1.note == Note.Cs3) {// Bass slap 1
-							return 1;
-						} else if (n2.note == Note.Cs3) {
-							return -1;
-						} else if (n1.note == Note.Cs4) {// Muted 2
-							return 1;
-						} else if (n2.note == Note.Cs4) {
-							return -1;
-						} else if (n1.note == Note.C3) {// Muted Mid
-							return 1;
-						} else if (n2.note == Note.C3) {
-							return -1;
-						}
-						
-						// Its too constrained to prioritize the rest.
-						// No way to really prioritize them, depends
-						// on song and transcribers taste.
-						//
-						// Note that muted 1 is not included on purpose.
 					}
-					
-					// discard the center-most note (for drum this is very random)
-					int center = Math.abs(n1.note.id - (lowest + (highest-lowest)/2))-Math.abs(n2.note.id - (lowest + (highest-lowest)/2));
-					
-					if (center > 0) return 1;
-					if (center < 0) return -1;
-					return 0;
-					
-					//1: n1 wins  -1: n2 wins   0:equal
+
+					if ((Math.abs(n1.note.id - n2.note.id) == 12 || Math.abs(n1.note.id - n2.note.id) == 24 || Math.abs(n1.note.id - n2.note.id) == 36)) {
+						// If 2 notes have octave spacing, keep the highest pitched.
+						if (n1.note.id > n2.note.id) {
+							points += 2;
+						} else if (n2.note.id > n1.note.id) {
+							points += -2;
+						}
+					}
+
+					if (sustained) {
+						if (n1.tiesFrom != null) {
+							points += 1;
+						}
+						if (n2.tiesFrom != null) {
+							points += -1;
+						}
+					}
+
+					if (points > 0) return 1;
+					if (points < 0) return -1;
+
+				} else {
+
+					// Bass drums get priority:
+					if (n1.note == Note.As3) {// Open bass
+						return 1;
+					} else if (n2.note == Note.As3) {
+						return -1;
+					} else if (n1.note == Note.D3) {// Bass slap 2
+						return 1;
+					} else if (n2.note == Note.D3) {
+						return -1;
+					} else if (n1.note == Note.Gs3) {// Bass
+						return 1;
+					} else if (n2.note == Note.Gs3) {
+						return -1;
+					} else if (n1.note == Note.Cs3) {// Bass slap 1
+						return 1;
+					} else if (n2.note == Note.Cs3) {
+						return -1;
+					} else if (n1.note == Note.Cs4) {// Muted 2
+						return 1;
+					} else if (n2.note == Note.Cs4) {
+						return -1;
+					} else if (n1.note == Note.C3) {// Muted Mid
+						return 1;
+					} else if (n2.note == Note.C3) {
+						return -1;
+					}
+
+					// Its too constrained to prioritize the rest.
+					// No way to really prioritize them, depends
+					// on song and transcribers taste.
+					//
+					// Note that muted 1 is not included on purpose.
 				}
+
+				// discard the center-most note (for drum this is very random)
+				int center = Math.abs(n1.note.id - (lowest + (highest-lowest)/2))-Math.abs(n2.note.id - (lowest + (highest-lowest)/2));
+
+				return Integer.compare(center, 0);
+
+				//1: n1 wins  -1: n2 wins   0:equal
 			};
 			
 			notes.sort(keepMe);
@@ -463,19 +457,17 @@ public class Chord implements AbcConstants
 		long endNoteTick = getStartTick(); 
 		if (!notes.isEmpty())
 		{
-			for (int k = 0; k < notes.size(); k++)
-			{
-				if (notes.get(k).note != Note.REST && notes.get(k).getEndTick() > endNoteTick)
-				{
-					endNoteTick = notes.get(k).getEndTick();
-				}
-			}
+            for (NoteEvent note : notes) {
+                if (note.note != Note.REST && note.getEndTick() > endNoteTick) {
+                    endNoteTick = note.getEndTick();
+                }
+            }
 		}
 		return endNoteTick;
 	}
 
 	public void removeRests() {
-		List<NoteEvent> rests = new ArrayList<NoteEvent>();
+		List<NoteEvent> rests = new ArrayList<>();
 		for (NoteEvent evt : notes) {
 			if (Note.REST == evt.note) {
 				rests.add(evt);
@@ -489,13 +481,11 @@ public class Chord implements AbcConstants
 		long endNoteTick = getEndTick(); 
 		if (!notes.isEmpty())
 		{
-			for (int k = 0; k < notes.size(); k++)
-			{
-				if (notes.get(k).note != Note.REST && notes.get(k).getEndTick() != endNoteTick)
-				{
-					System.out.println("Note in chord has bad length! "+(notes.get(k).getEndTick() - endNoteTick));
-				}
-			}
+            for (NoteEvent note : notes) {
+                if (note.note != Note.REST && note.getEndTick() != endNoteTick) {
+                    System.out.println("Note in chord has bad length! " + (note.getEndTick() - endNoteTick));
+                }
+            }
 		}		
 	}
 }
