@@ -38,22 +38,22 @@ public enum LotroInstrument
 	BASIC_HARP               ( "Basic Harp",                false, MidiInstrument.ORCHESTRA_HARP,       0,      false,     6.0f, "Harp", "harfe", "harpe", "harpe (de)? base"),
 	MISTY_MOUNTAIN_HARP      ( "Misty Mountain Harp",       false, MidiInstrument.CLEAN_ELEC_GUITAR,    0,      false,   -12.5f, "Misty Harp", "MM Harp", "MMH", "Harpe (des)? monts brumeux"),
 	BASIC_LUTE               ( "Basic Lute",                false, MidiInstrument.STEEL_STRING_GUITAR,  0,      false,   -19.0f, "New Lute", "LuteB", "Banjo", "Luth (de)? base"),
-	LUTE_OF_AGES             ( "Lute of Ages",              false, MidiInstrument.NYLON_GUITAR,         0,      false,     0.0f, "Lute", "Age Lute", "LuteA", "LOA", "Guitar", "laute", "luth", "luth (des)? si�cles"),
-	BASIC_THEORBO            ( "Basic Theorbo",             false, MidiInstrument.ACOUSTIC_BASS,       -1,      false,   -12.0f, "Theorbo", "Theo", "Bass", "theorbe", "th�orbe"),
+	LUTE_OF_AGES             ( "Lute of Ages",              false, MidiInstrument.NYLON_GUITAR,         0,      false,     0.0f, "Lute", "Age Lute", "LuteA", "LOA", "Guitar", "laute", "luth", "luth (des)? siècles"),
+	BASIC_THEORBO            ( "Basic Theorbo",             false, MidiInstrument.ACOUSTIC_BASS,       -1,      false,   -12.0f, "Theorbo", "Theo", "Bass", "theorbe", "théorbe"),
 	TRAVELLERS_TRUSTY_FIDDLE ( "Traveller's Trusty Fiddle", false, MidiInstrument.PIZZICATO_STRINGS,    1,      false,    -3.0f, "Travell?er'?s? (Trusty)? Fiddle", "Trusty Fiddle", "TT Fiddle", "violon (de)? voyageur"),
 
 	BARDIC_FIDDLE            ( "Bardic Fiddle",              true, MidiInstrument.VIOLIN,               1,      false,     6.0f, "Fiddle", "Violin", "Barden-?Geige", "violon (de)? barde"),
 	BASIC_FIDDLE             ( "Basic Fiddle",               true, MidiInstrument.VIOLA,                1,      false,     6.25f, "Violon (de)? base"),
 	LONELY_MOUNTAIN_FIDDLE   ( "Lonely Mountain Fiddle",     true, MidiInstrument.SYNTH_STRING_2,       1,      false,     5.5f, "Lonely Fiddle", "LM Fiddle", "LM fidel", "LM geige", "Violon (du)? Mont solitaire"),
 	SPRIGHTLY_FIDDLE         ( "Sprightly Fiddle",          false, MidiInstrument.FIDDLE,               1,      false,   -10.0f, "Muntere Geige", "Violon alerte"),
-	STUDENT_FIDDLE           ( "Student's Fiddle",           true, MidiInstrument.GUITAR_FRET_NOISE,    1,      false,     0.0f, "Student'?s? Fiddle", "Violon d'�tudes"),
+	STUDENT_FIDDLE           ( "Student's Fiddle",           true, MidiInstrument.GUITAR_FRET_NOISE,    1,      false,     0.0f, "Student'?s? Fiddle", "Violon d'études"),
 
 	BASIC_BAGPIPE            ( "Basic Bagpipe",              true, MidiInstrument.BAG_PIPE,             1,      false,    -1.5f, "Bag pipes?", "Pipes?", "dudelsack", "sack", "cornemuse", "cornemuse (de)? base"),
 	BASIC_BASSOON            ( "Basic Bassoon",              true, MidiInstrument.BASSOON,              0,      false,     5.0f, "Bassoon"),
 	BRUSQUE_BASSOON          ( "Brusque Bassoon",           false, MidiInstrument.OBOE,                 0,      false,     5.0f, "Brusk Bassoon", "Schroffes Fagott"),
 	LONELY_MOUNTAIN_BASSOON  ( "Lonely Mountain Bassoon",    true, MidiInstrument.SYNTH_BRASS_2,        0,      false,     5.0f, "Lonely Bassoon", "LM Bassoon", "Fagott (vom)? Einsamen ?Berg", "fagott", "Basson (du)? Mont Solitaire"),
 	BASIC_CLARINET           ( "Basic Clarinet",             true, MidiInstrument.CLARINET,             1,      false,    -2.0f, "Clarinet", "Clari", "klarinette", "clarinette", "clarinette (de)? base"),
-	BASIC_FLUTE              ( "Basic Flute",                true, MidiInstrument.FLUTE,                2,      false,    -3.5f, "Flute", "fl�te", "fl�t", "fl�t (de)? base"),
+	BASIC_FLUTE              ( "Basic Flute",                true, MidiInstrument.FLUTE,                2,      false,    -3.5f, "Flute", "flöte", "flût", "flût (de)? base"),
 	BASIC_HORN               ( "Basic Horn",                 true, MidiInstrument.ENGLISH_HORN,         0,      false,    -2.0f, "Horn", "cor", "cor (de)? base"),
 	BASIC_PIBGORN            ( "Basic Pibgorn",              true, MidiInstrument.CHARANG,              2,      false,    -3.5f, "Pibgorn"),
 
@@ -80,7 +80,7 @@ public enum LotroInstrument
 	public final float dBVolumeAdjust;
 	private final String[] nicknameRegexes;
 
-	private LotroInstrument(String friendlyName, boolean sustainable, MidiInstrument midiInstrument, int octaveDelta,
+	LotroInstrument(String friendlyName, boolean sustainable, MidiInstrument midiInstrument, int octaveDelta,
 			boolean isPercussion, float dBVolumeAdjust, String... nicknameRegexes)
 	{
 		this.lowestPlayable = Note.MIN_PLAYABLE;
@@ -162,7 +162,7 @@ public enum LotroInstrument
 			}
 		}
 
-		return new Pair<LotroInstrument, MatchResult>(instrument, result);
+		return new Pair<>(instrument, result);
 	}
 
 	public static LotroInstrument findInstrumentName(String str, LotroInstrument defaultInstrument)
