@@ -5,9 +5,9 @@ package com.digero.common.util;
  */
 public interface ICompileConstants
 {
-	public static final boolean MUTE_DISABLED_TRACKS = false;
-	public static final boolean SHOW_TEMPO_SPINNER = true;
-	public static final boolean SHOW_METER_TEXTBOX = true;
-	public static final boolean SHOW_KEY_FIELD = false;
-	public static final boolean SHOW_COLORIZER = false;
+	boolean MUTE_DISABLED_TRACKS = false;
+	boolean SHOW_TEMPO_SPINNER = true;
+	boolean SHOW_METER_TEXTBOX = true;
+	boolean SHOW_KEY_FIELD = false;
+	boolean SHOW_COLORIZER = false;
 }
