@@ -2,5 +2,5 @@ package com.digero.common.util;
 
 public interface IDiscardable
 {
-	public void discard();
+	void discard();
 }
