@@ -12,6 +12,7 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.UIManager;
 
+import com.digero.common.midi.MidiUtils;
 import com.digero.common.midi.Note;
 import com.digero.common.midi.SequencerEvent;
 import com.digero.common.midi.SequencerEvent.SequencerProperty;
@@ -25,7 +26,6 @@ import com.digero.maestro.midi.SequenceDataCache;
 import com.digero.maestro.midi.SequenceDataCache.TempoEvent;
 import com.digero.maestro.midi.SequenceInfo;
 import com.digero.maestro.view.TrackPanel.TrackDimensions;
-import com.sun.media.sound.MidiUtils;
 
 @SuppressWarnings("serial")
 public class TempoPanel extends JPanel implements IDiscardable, TableLayoutConstants
@@ -201,7 +201,7 @@ public class TempoPanel extends JPanel implements IDiscardable, TableLayoutConst
 		private void recalcTempoEvents()
 		{
 			// Make fake note events for every tempo event
-			events = new ArrayList<NoteEvent>();
+			events = new ArrayList<>();
 			TempoEvent prevEvent = null;
 			SequenceDataCache dataCache = sequenceInfo.getDataCache();
 			for (TempoEvent event : dataCache.getTempoEvents().values())
